Sprint1 unittest

See merge request 2ti_project_informatica_2020_2021/groep-5/leap-groep-5!7

+++ backend/src/main/java/com/bavostepbros/leap/domain/service/StrategyService/StrategyServiceImpl.java
@@ -8,19 +8,25 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.transaction.Transactional;
 
+import com.bavostepbros.leap.domain.customexceptions.DuplicateValueException;
+import com.bavostepbros.leap.domain.customexceptions.ForeignKeyException;
+import com.bavostepbros.leap.domain.customexceptions.IndexDoesNotExistException;
+import com.bavostepbros.leap.domain.customexceptions.InvalidInputException;
+import com.bavostepbros.leap.domain.customexceptions.StrategyException;
 import com.bavostepbros.leap.domain.model.Environment;
 import com.bavostepbros.leap.domain.model.Status;
 import com.bavostepbros.leap.domain.model.Strategy;
-import com.bavostepbros.leap.persistence.EnvironmentDAL;
+import com.bavostepbros.leap.domain.service.environmentservice.EnvironmentService;
+import com.bavostepbros.leap.domain.service.statusservice.StatusService;
 import com.bavostepbros.leap.persistence.StrategyDAL;
 
 import lombok.RequiredArgsConstructor;
 
 /**
-*
-* @author Lenny Bontenakel, Bavo Van Meel
-*
-*/
+ *
+ * @author Lenny Bontenakel, Bavo Van Meel
+ *
+ */
 @Service
 @Transactional
 @RequiredArgsConstructor
@@ -28,15 +34,37 @@ public class StrategyServiceImpl implements StrategyService {
 
 	@Autowired
 	private StrategyDAL strategyDAL;
-	
+
 	@Autowired
-	private EnvironmentDAL environmentDAL;
+	private EnvironmentService environmentService;
+
+	@Autowired
+	private StatusService statusService;
 
 	@Override
-	public Strategy save(Integer statusId, LocalDate validityPeriod, String strategyName, LocalDate timeFrameStart,
-			LocalDate timeFrameEnd, Integer environmentId, String environmentName) {
-		Status status = new Status(statusId, validityPeriod);
-		Environment environment = new Environment(environmentId, environmentName);
+	public Strategy save(Integer statusId, String strategyName, LocalDate timeFrameStart, LocalDate timeFrameEnd,
+			Integer environmentId) {
+		if (strategyName == null || strategyName.isBlank() || strategyName.isEmpty()) {
+    		throw new InvalidInputException("Invalid input.");
+    	}
+		if (statusId == null || statusId.equals(0)) {
+			throw new ForeignKeyException("Status ID is invalid.");
+		}		
+		if (environmentId == null || environmentId.equals(0)) {
+			throw new ForeignKeyException("Environment ID is invalid.");
+		}		
+    	if (!existsByStrategyName(strategyName)) {
+			throw new DuplicateValueException("Strategy name already exists.");
+		}
+    	if (!statusService.existsById(statusId)) {
+			throw new ForeignKeyException("Status ID does not exists.");
+		}
+		if (!environmentService.existsById(environmentId)) {
+			throw new ForeignKeyException("Environment ID does not exists.");
+		}
+    	
+		Status status = statusService.get(statusId);
+		Environment environment = environmentService.get(environmentId);
 		Strategy strategy = new Strategy(status, strategyName, timeFrameStart, timeFrameEnd, environment);
 		Strategy savedStrategy = strategyDAL.save(strategy);
 		return savedStrategy;
@@ -44,6 +72,13 @@ public class StrategyServiceImpl implements StrategyService {
 
 	@Override
 	public Strategy get(Integer id) {
+		if (id == null || id.equals(0)) {
+			throw new InvalidInputException("Strategy ID is not valid.");
+		}
+		if (!existsById(id)) {
+			throw new IndexDoesNotExistException("Strategy ID does not exists.");
+		}
+		
 		Strategy strategy = strategyDAL.findById(id).get();
 		return strategy;
 	}
@@ -55,10 +90,27 @@ public class StrategyServiceImpl implements StrategyService {
 	}
 
 	@Override
-	public Strategy update(Integer strategyId, Integer statusId, LocalDate validityPeriod, String strategyName,
-			LocalDate timeFrameStart, LocalDate timeFrameEnd, Integer environmentId, String environmentName) {
-		Status status = new Status(statusId, validityPeriod);
-		Environment environment = new Environment(environmentId, environmentName);
+	public Strategy update(Integer strategyId, Integer statusId, String strategyName, LocalDate timeFrameStart,
+			LocalDate timeFrameEnd, Integer environmentId) {
+		if (strategyId == null || strategyId.equals(0) || strategyName == null || 
+    			strategyName.isBlank() || strategyName.isEmpty()) {
+			throw new InvalidInputException("Invalid input.");
+		}
+		if (!existsById(strategyId)) {
+			throw new StrategyException("Can not update strategy if it does not exist.");
+		}
+		if (!existsByStrategyName(strategyName)) {
+			throw new DuplicateValueException("Strategy name already exists.");
+		}
+		if (!statusService.existsById(statusId)) {
+			throw new ForeignKeyException("Status ID does not exists.");
+		}
+		if (!environmentService.existsById(environmentId)) {
+			throw new ForeignKeyException("Environment ID does not exists.");
+		}
+		
+		Status status = statusService.get(statusId);
+		Environment environment = environmentService.get(environmentId);
 		Strategy strategy = new Strategy(strategyId, status, strategyName, timeFrameStart, timeFrameEnd, environment);
 		Strategy updatedStrategy = strategyDAL.save(strategy);
 		return updatedStrategy;
@@ -66,24 +118,35 @@ public class StrategyServiceImpl implements StrategyService {
 
 	@Override
 	public void delete(Integer id) {
+		if (id == null || id.equals(0)) {
+			throw new InvalidInputException("Strategy ID is not valid.");
+		}
+		if (!existsById(id)) {
+			throw new IndexDoesNotExistException("Strategy ID does not exists.");
+		}
 		strategyDAL.deleteById(id);
 	}
 
 	@Override
 	public boolean existsById(Integer id) {
-		boolean result = strategyDAL.existsById(id);
-		return result;
+		return strategyDAL.existsById(id);
 	}
 
 	@Override
 	public boolean existsByStrategyName(String strategyName) {
-		boolean result = strategyDAL.findByStrategyName(strategyName).isEmpty();
-		return result;
+		return strategyDAL.findByStrategyName(strategyName).isEmpty();
 	}
 
 	@Override
 	public List<Strategy> getStrategiesByEnvironment(Integer environmentId) {
-		Environment environment = environmentDAL.findById(environmentId).get();
+		if (environmentId == null || environmentId.equals(0)) {
+			throw new InvalidInputException("Environment ID is not valid.");
+		}
+		if (!environmentService.existsById(environmentId)) {
+			throw new ForeignKeyException("Environment ID does not exists.");
+		}
+		
+		Environment environment = environmentService.get(environmentId);
 		List<Strategy> strategies = strategyDAL.findByEnvironment(environment);
 		return strategies;
 	}
